# backend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "TarotGen.io API",
        "description": "The backend for **TarotGen.io** — a PHP 8.5 / Slim 4 hybrid-REST API.\n\n- Resources are plural nouns (`/decks`, `/readings`); items are `/{id}`.\n- `GET` reads, `POST` creates (→`201`) or runs a command, `PUT` replaces,\n  `PATCH` applies a partial update / flag toggle, `DELETE` removes (→`204`).\n- State transitions that aren't a plain field-set are `POST …/{id}/{verb}`\n  (e.g. `…/finalize`, `…/approve`, `…/draw`).\n- Errors are `{ \"error\": \"…\" }`; multi-message validation is `{ \"errors\": […] }`.\n- Auth is a session cookie (`PHPSESSID`) set by the `/auth/*` endpoints —\n  there are no API keys or bearer tokens for first-party calls.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Same-origin API base path"
        }
    ],
    "paths": {
...
        "/auth/google": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Redirect to Google's consent screen",
                "description": "Initiate the OAuth flow. Accepts an optional ?intent query param\n('login', 'register', or 'link') so the callback knows what to do.",
                "operationId": "eac6533cfebab1e8ffaeaaadc83f09d3",
                "parameters": [
                    {
                        "name": "intent",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "login",
                                "register",
                                "link"
                            ]
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect to Google"
                    },
                    "401": {
                        "description": "The link intent requires a session"
                    },
                    "503": {
                        "description": "Google OAuth is not configured"
                    }
                }
            }
        },
        "/auth/google/callback": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "OAuth callback — redirects back into the SPA on success/failure",
                "description": "Handle the OAuth callback from Google.",
                "operationId": "735cd3cd48a4b0bc2f174da448838cf8",
                "parameters": [
                    {
                        "name": "code",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "state",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect into the SPA"
                    }
                }
            }
        },
        "/auth/google/unlink": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Unlink Google from the current account",
                "description": "Unlink Google from the currently logged-in account.",
                "operationId": "f8948158ac3bd6844a2eae03c72c562d",
                "responses": {
                    "200": {
                        "description": "Unlinked",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Account has no password set"
                    },
                    "401": {
                        "description": "Not authenticated"
                    }
                },
                "security": [
                    {
                        "sessionCookie": []
                    }
                ]
            }
        },
...
        "/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Create an account (inactive until email activation). Rate-limited: 5/IP/hour.",
                "operationId": "ad99e4939452f71308786370a3ed004a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "display_name",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "display_name": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "minLength": 12
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Account created"
                    },
                    "422": {
                        "description": "Validation/uniqueness failure"
                    },
                    "429": {
                        "description": "Rate limit exceeded"
                    }
                }
            }
        },
...
        "/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Password sign-in. Rate-limited: 10 failed attempts/IP/15 min.",
                "operationId": "ba29841b2cc01b0ea43c8ad6b1b529b3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "remember_me": {
                                        "type": "boolean"
                                    },
                                    "turnstile_token": {
                                        "description": "Required when Turnstile is configured",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Signed in; sets the session cookie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Captcha failed"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    },
                    "403": {
                        "description": "Inactive or password-login disabled"
                    },
                    "429": {
                        "description": "Rate limit exceeded"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "User": {
                "description": "A user account in its safe shape (no secrets or hashes).",
                "properties": {
                    "user_id": {
                        "type": "integer"
                    },
                    "email": {
                        "type": "string"
                    },
                    "display_name": {
                        "type": "string"
                    },
                    "is_active": {
                        "type": "boolean"
                    },
                    "is_admin": {
                        "type": "boolean"
                    },
                    "registered_at": {
                        "type": "string"
                    },
                    "last_login_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "google_linked": {
                        "type": "boolean"
                    },
                    "password_login_disabled": {
                        "type": "boolean"
                    },
                    "has_passkeys": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sessionCookie": {
                "type": "apiKey",
                "description": "Session cookie issued on sign-in. Sent automatically by the browser.",
                "name": "PHPSESSID",
                "in": "cookie"
            }
        }
    },
...
}